reuse score schema/type
